feat: redirect custom hook 사용 위치 변경, 로그인 정보에 상관없이 메인으로 이동
